requirements.txt
absl_py==2.3.1
accelerate==1.9.0
accelerated_scan==0.2.0
addict==2.4.0
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.4.0
alabaster==1.0.0
alembic==1.16.4
aniso8601==10.0.1
annotated_types==0.7.0
antlr4_python3_runtime==4.9.3
anyio==4.9.0
asciitree==0.3.3
asgiref==3.9.1
asttokens==3.0.0
attrdict==2.0.1
attrs==25.3.0
audioread==3.0.1
av==15.0.0
babel==2.17.0
backoff==2.2.1
bcrypt==4.3.0
beautifulsoup4==4.13.1
black==24.10.0
blinker==1.9.0
boto3==1.39.12
botocore==1.39.12
braceexpand==0.1.7
bracex==2.6
build==1.2.2.post1
cached_property==2.0.1
cachetools==5.5.2
cdifflib==1.2.6
certifi==2025.7.14
cffi==1.17.1
chardet==5.2.0
charset_normalizer==3.4.2
chex==0.1.89
chromadb==1.0.15
click==8.2.1
clip==0.2.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.9.0
contourpy==1.3.2
coverage==7.9.2
cryptography==45.0.4
cycler==0.12.1
cython==3.1.2
cytoolz==1.0.1
dataclasses_json==0.6.7
DataProperty==1.1.0
datasets==2.19.0
decorator==5.2.1
Deprecated==1.2.18
diffusers==0.34.0
dill==0.3.8
distance==0.1.3
distro==1.9.0
django_auditlog==3.2.0
django_filter==25.1
django_js_asset==3.1.2
django_jsonform==2.23.2
django_mptt==0.14.0
django==4.2.23
djangorestframework==3.16.0
docopt==0.6.2
docutils==0.21.2
drf_excel==2.5.3
drf_flex_fields==1.0.2
durationpy==0.10
editdistance==0.8.1
einops_exts==0.0.4
einops==0.8.1
emoji==2.14.1
et_xmlfile==2.0.0
etils==1.13.0
evaluate==0.4.5
executing==2.2.0
faiss_cpu==1.11.0.post1
fastapi==0.116.1
fasteners==0.19
fiddle==0.3.0
filelock==3.18.0
filetype==1.2.0
Flask_RESTful==0.3.10
flask==3.1.1
flatbuffers==25.2.10
flax==0.10.7
fonttools==4.59.0
frozenlist==1.7.0
fsspec==2024.3.1
ftfy==6.3.1
future==1.0.0
g2p_en==2.1.0
gdown==5.2.0
gitdb==4.0.12
gitpython==3.1.45
google_api_core==2.25.1
google_api_python_client==2.177.0
google_auth_httplib2==0.2.0
google_auth==2.40.3
googleapis_common_protos==1.70.0
graphviz==0.21
grpcio==1.73.1
h11==0.16.0
h5py==3.14.0
hf_xet==1.1.5
html2text==2025.4.15
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx_sse==0.4.1
httpx==0.28.1
huggingface_hub==0.21.3
humanfriendly==10.0
humanize==4.12.3
hydra_core==1.3.2
hypothesis==6.135.32
idna==3.10
ijson==3.4.0
imageio==2.37.0
imagesize==1.4.1
immutabledict==4.2.0
importlib_metadata==8.7.0
importlib_resources==6.5.2
inflect==7.5.0
iniconfig==2.1.0
intervaltree==3.1.0
ipython_pygments_lexers==1.1.1
ipython==9.4.0
isort==5.13.2
itsdangerous==2.2.0
Janome==0.5.0
jax==0.4.38
jaxlib==0.4.38
jedi==0.19.2
jieba==0.42.1
jinja2==3.1.6
jiter==0.10.0
jiwer==3.1.0
jmespath==1.0.1
joblib==1.5.1
jsonlines==4.0.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_specifications==2025.4.1
jsonschema==4.25.0
kaldi_io==0.9.8
kaldi_python_io==1.2.2
kaldiio==2.18.1
kiwisolver==1.4.8
kornia_rs==0.1.9
kornia==0.8.1
kubernetes==33.1.0
langchain_community==0.3.27
langchain_core==0.3.72
langchain_openai==0.3.28
langchain_text_splitters==0.3.9
langchain==0.3.27
langdetect==1.0.9
langgraph_checkpoint==2.1.1
langgraph_prebuilt==0.6.2
langgraph_sdk==0.2.0
langgraph==0.6.2
langsmith==0.4.8
lark==1.2.2
latexcodec==3.0.1
lazy_loader==0.4
ldap3==2.9.1
levenshtein==0.27.1
lhotse==1.30.3
libcst==1.8.2
librosa==0.11.0
lightning_utilities==0.14.3
lightning==2.4.0
lilcom==1.8.1
llvmlite==0.44.0
logging==0.4.9.6
loguru==0.7.3
lxml==6.0.0
mako==1.3.10
markdown_it_py==3.0.0
markdown==3.8.2
markdown2==2.5.3
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib_inline==0.1.7
matplotlib==3.10.3
mbstrdecoder==1.1.4
mdurl==0.1.2
mediapy==1.1.6
megatron_core==0.12.2
megatron_energon==5.2.0
ml_dtypes==0.5.1
mmh3==5.2.0
more_itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
multi_storage_client==0.25.0
multidict==6.6.3
multiprocess==0.70.16
mypy_extensions==1.1.0
nemo_toolkit==1.23.0rc0
nemo==7.2.3
nerfacc==0.5.3
nest_asyncio==1.6.0
networkx==3.5
nltk==3.9.1
num2words==0.5.14
numba==0.61.0
numcodecs==0.15.1
numexpr==2.11.0
numpy==1.26.4
nvidia_lm_eval==25.6.1
oauthlib==3.3.1
omegaconf==2.3.0
onnx==1.17.0
onnxruntime==1.22.1
open_clip_torch==2.24.0
openai==1.97.1
opencc==1.1.9
openpyxl==3.1.5
opentelemetry_api==1.36.0
opentelemetry_exporter_otlp_proto_common==1.36.0
opentelemetry_exporter_otlp_proto_grpc==1.36.0
opentelemetry_proto==1.36.0
opentelemetry_sdk==1.36.0
opentelemetry_semantic_conventions==0.57b0
opt_einsum==3.4.0
optax==0.2.5
optuna==4.4.0
orbax_checkpoint==0.11.19
orjson==3.11.1
ormsgpack==1.10.0
overrides==7.7.0
packaging==24.2
pandas==2.3.1
pangu==4.0.6.1
parameterized==0.9.0
parso==0.8.4
pathspec==0.12.1
pathvalidate==3.3.1
peft==0.16.0
pesq==0.0.4
pexpect==4.9.0
pillow==11.2.1
pip==25.1.1
plac==1.4.5
platformdirs==4.3.8
pluggy==1.6.0
pooch==1.8.2
portalocker==3.2.0
posthog==5.4.0
prettytable==3.16.0
progress==1.6.1
prompt_toolkit==3.0.51
propcache==0.3.2
proto_plus==1.26.1
protobuf==6.31.1
psutil==7.0.0
psycopg2_binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.3
pyannote.metrics==3.2.1
pyarrow_hotfix==0.7
pyarrow==20.0.0
pyasn1_modules==0.4.2
pyasn1==0.6.1
pybase64==1.4.2
pybind11==3.0.0
pybtex_docutils==1.0.3
pybtex==0.25.1
pycparser==2.22
pyctcdecode==0.5.0
pydantic_core==2.33.2
pydantic_settings==2.10.1
pydantic==2.11.7
pydub==0.25.1
pygments==2.19.2
pygtrie==2.5.0
pyloudnorm==0.1.1
pymodbus==3.8.6
pyparsing==3.2.3
pypdf==5.8.0
pypika==0.48.9
pypinyin_dict==0.9.0
pypinyin==0.55.0
pyproject_hooks==1.2.0
PySocks==1.7.1
pystoi==0.4.1
pytablewriter==1.2.1
pytest_cov==6.2.1
pytest_mock==3.14.1
pytest_random_order==1.2.0
pytest_runner==6.0.1
pytest==8.4.1
python_dateutil==2.9.0
python_dotenv==1.1.1
python_iso639==2025.2.18
python_magic==0.4.27
pytorch_lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
qwen_vl_utils==0.0.11
rapidfuzz==3.13.0
referencing==0.36.2
regex==2024.11.6
requests_oauthlib==2.0.0
requests_toolbelt==1.0.0
requests==2.32.3
resampy==0.4.3
rich==14.0.0
roman_numerals_py==3.1.0
rouge_score==0.1.2
rpds_py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
s3fs==0.4.2
s3transfer==0.13.1
sacrebleu==2.5.1
sacremoses==0.1.1
safetensors==0.5.3
sarvamai==0.1.15
scikit_learn==1.7.0
scipy==1.16.0
seaborn==0.13.2
sentence_transformers==5.0.0
sentencepiece==0.2.0
sentry_sdk==2.33.2
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
snowballstemmer==3.0.1
sortedcontainers==2.4.0
soundfile==0.13.1
soupsieve==2.7
sox==1.5.0
soxr==0.5.0.post1
sphinx==8.2.3
sphinxcontrib_applehelp==2.0.0
sphinxcontrib_bibtex==2.6.5
sphinxcontrib_devhelp==2.0.0
sphinxcontrib_htmlhelp==2.1.0
sphinxcontrib_jsmath==1.0.1
sphinxcontrib_qthelp==2.0.0
sphinxcontrib_serializinghtml==2.0.0
sqlalchemy==2.0.41
sqlparse==0.5.3
stack_data==0.6.3
starlette==0.47.2
sympy==1.14.0
tabledata==1.3.4
tabulate==0.9.0
taming_transformers==0.0.1
tcolorpy==0.1.7
tenacity==9.1.2
tensorboard_data_server==0.7.2
tensorboard==2.19.0
tensorstore==0.1.76
termcolor==3.1.0
text_unidecode==1.3
textdistance==4.6.3
texterrors==0.5.1
threadpoolctl==3.6.0
tiktoken==0.7.0
timm==1.0.19
tokenizers==0.19.0
toolz==1.0.0
torch_stft==0.1.4
torch==2.7.1
torchaudio==2.7.1
torchdiffeq==0.2.5
torchmetrics==1.8.0
torchsde==0.2.6
torchvision==0.22.1
tqdm_multiprocess==0.0.11
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformers==4.40.0
tree_sitter_python==0.23.6
tree_sitter==0.25.0
treescope==0.1.9
trimesh==4.7.1
typeguard==4.4.4
typepy==1.3.4
typer==0.16.0
typing_extensions==4.14.1
typing_inspect==0.9.0
typing_inspection==0.4.1
tzdata==2025.2
Unidecode==1.4.0
unstructured_client==0.40.0
unstructured==0.14.9
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.21.0
vapi_server_sdk==1.7.0
wandb==0.21.0
watchfiles==1.1.0
wcmatch==10.1
wcwidth==0.2.13
webdataset==1.0.2
websocket_client==1.8.0
websockets==15.0.1
werkzeug==3.1.3
wget==3.2
whisper_jax==0.0.1
word2number==1.1
wrapt==1.17.2
xattr==1.1.4
xxhash==3.5.0
yarl==1.20.1
zarr==2.18.7
zipp==3.23.0
zstandard==0.23.0
